Upgrade to SilverStripe 4
