+++ README.md
@@ -7,7 +7,7 @@ Send CMS managed system email notifications from code.
 *  Shea Dawson (<shea@symbiote.com.au>)
 
 ## Requirements
-* SilverStripe 3.1 +
+* SilverStripe 4.0 +
 
 ## Installation Instructions
 
@@ -21,7 +21,7 @@ composer require symbiote/silverstripe-notifications
 In your _config yml file, add an identifier for each notification you require. This allows you to lookup Notification objects in the database from your code. 
 
 ```
-SystemNotification:
+Symbiote\Notifications\Model\SystemNotification:
   identifiers:
     - 'NAME_OF_NOTIFICATION1'
     - 'NAME_OF_NOTIFICATION2'
@@ -31,6 +31,8 @@ SystemNotification:
 Add the NotifiedOn interface to any dataobjects that are relevant to the notifications you will be sending. This is required so the Notifications module can look up the below methods on your object to send the notification.
 
 ```php
+use Symbiote\Notifications\Model\NotifiedOn;
+
 class MyDataObject extends DataObject implements NotifiedOn {
 	...
 ```
@@ -56,7 +58,7 @@ public function getNotificationTemplateData();
 
 Note: the follow template data is automatically included:
 
-* $ThemeDir
+* $ThemeDirs (an ArrayList object of themes, if you only have one theme using `$ThemeDirs.First` should be the same as the old `$ThemeDir` )
 * $SiteConfig
 * $MyDataObject (whatever the ClassName of your NotifiedOn DataObject is)
 * $Member (The Member object this message is being sent to)
@@ -66,7 +68,7 @@ Note: the follow template data is automatically included:
  * Gets the list of recipients for a given notification event, based on this object's 
  * state. 
  * $event The identifier of the event that triggered this notification
- * @return array An array of Member objects	
+ * @return array An array of Member objects
  */
 public function getRecipients($event);
 ```
@@ -80,7 +82,9 @@ Create a notification in the Notifications model admin, in the CMS.
 ### 4)
 Send the notification from your code, where $contextObject is an instance of the DataObject being NotifiedOn 
 ```php
-singleton('NotificationService')->notify('NOTIFICATION_IDENTIFIER', $contextObject);
+use Symbiote\Notifications\Service\NotificationService;
+
+singleton(NotificationService::class)->notify('NOTIFICATION_IDENTIFIER', $contextObject);
 ```
 
 ## Templates
@@ -88,7 +92,7 @@ singleton('NotificationService')->notify('NOTIFICATION_IDENTIFIER', $contextObje
 Notifications can be rendered with .ss templates. This is useful if you want to have a header/footer in your email notifications. You can either specify a template on a per/notification basis in the CMS, and/or set a default template for all notifications to be rendered with:
 
 ```
-SystemNotification:
+Symbiote\Notifications\Model\SystemNotification:
   default_template: EmailNotification
 ```
 
@@ -98,7 +102,7 @@ In your templates, you render the notification text with the $Body variable.
 
 You will probably want to configure a send_from email address - 
 ```
-EmailNotificationSender:
+Symbiote\Notifications\Service\EmailNotificationSender:
   send_notifications_from: 'notifications@example.com'
 ```
