+++ tests/NotificationsTest.yml
@@ -1,4 +1,4 @@
-NotifyOnThis:
+Symbiote\Notifications\Tests\NotifyOnThis:
   not1:
     Title: Some Data Object
     Status: Available
